supercadernos_v5.0/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "supercadernos"
version = "5.0.0"
description = "Gerador de Supercadernos — v5.0 (CLI, cache JSON, pré-filtro, logging, resume)"
authors = [{name="Time Supercadernos"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.1",
  "jsonschema>=4.23.0",
  "pypdf>=5.0.0",
  "PyMuPDF>=1.24.9",
  "rich>=13.7.1",
  "unidecode>=1.3.8"
]

[project.scripts]
supercadernos = "supercadernos.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
